pyproject.toml
[project]
name = "serper-mcp-server"
version = "0.1.0"
description = "Add your description here"
keywords = ["serper", "mcp", "server", "google", "search"]
authors = [
    { name="Gary Meng", email="guozao.m@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[project.scripts]
serper-mcp-server = "serper_mcp_server:main"